requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
alpaca-py>=0.21.0
python-dotenv>=1.0.0
